requirements.txt
cx_Oracle==8.3.0
duckdb==0.9.2
python-dotenv==1.0.0